+++ sound-mind-studio/tests/test_canvas_widget.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <QObject>
+
+class CanvasWidgetTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void sizeHintFallsBackWithNoProject();
+    void sizeHintMatchesProjectCanvasDimensions();
+};