+++ sound-mind-studio/tests/test_main_window.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <QObject>
+
+class MainWindowTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void startsWithAFreshProject();
+    void newProjectReplacesTheCurrentOne();
+};